feature: Initial implementation (#1)
This initial implementation generates reports of the current assets in
the inventory and also closed positions (assets that were in the
inventory on the last year and are not present anymore).

Future improvements will take care of parsing buys and sells on the B3
for figuring out the invested value for those positions.

Signed-off-by: Filipe Utzig <filipe@gringolito.com>